Fix state machine transision bug from PR #304
User operation will not be able to continue if the validation failed during
the publish, since the logic to interrupt publish when validation failed is
taking place in the beginning of extraction.

In the state machine, `extracting` state will need to transition to `finished`
when the validation failed, and since it's not able to do so, operation
can't continue, due to the stop button didn't change to reset button.

This commit fixed that.