Change logic to get out of the iterator
When no error message needs to be passed, use 'return' to exit the
generator as recommended in PEP-479.

When an error message needs to be passed, raise a RuntimeError exception
instead to ease the data flow. No need to catch exception in the
iterator as we will deal globally with exceptions raised at higher
level.