Remove vendorized mock.py
The vendorized version of the mock module has been unsupported for a long time and has Python 3 issues. The module, however, does not appear to be used anywhere so those issues aren't being surfaced. Rather than fix or continue ignoring them I'm removing the module. If needed a more recent version of the module was added to the standard library at unittest.mock in Python 3.3.